Move Territory model into separate file
